Add support for MCP3004, MCP3204 and MCP3208
The MCP3004 and MCP3008 have almost identical implementation, as well as
the MCP3204 and MCP3208. Although differences between MCP3004/8 vs
MCP3204/8 are small, it was hard to come up with a shared abstraction
that wasn't very hard to read and understant. Therefore I wrote 2
implementations which lead to some duplicate code.