Split adc package in io/adc and io/spi/microchip
The ADC interface and related constants have been moved to a
seperate package called io/adc. The ADC interface has been changed.
A method ADC.OutputCode(channel int) (int, error) hass been added.
This method returns the digital output code of a channel. This
is the raw value that one gets when querying a channel.
It is a integer between 0 and the maximum resolution of the ADC.
To reflect this change the ADC.Read(channel int) (float64, error)
has been renamed to ADC.Voltage(channel int) (float64, error).

The logic for the MCP3004, MCP3008, MCP3204 and MCP3208 has been
moved to the package io/spi/microchip.